+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/odata/retriever/OrganizationODataRetriever.java
@@ -103,7 +103,7 @@ public class OrganizationODataRetriever
 	}
 
 	@Override
-	public long getResultsCount(
+	public int getResultsCount(
 			long companyId, String filterString, Locale locale)
 		throws PortalException {
 
@@ -132,7 +132,8 @@ public class OrganizationODataRetriever
 				return hits.getLength();
 			}
 
-			return IndexSearcherHelperUtil.searchCount(searchContext1, query);
+			return
+				(int)IndexSearcherHelperUtil.searchCount(searchContext1, query);
 		}
 		catch (Exception e) {
 			throw new PortalException(
